+++ test/Fixture/Definitions/ImplementsDefinitionButHasPrivateConstructor/RepositoryDefinition.php
@@ -16,7 +16,6 @@ namespace Ergebnis\FactoryBot\Test\Fixture\Definitions\ImplementsDefinitionButHa
 use Ergebnis\FactoryBot\Definition;
 use Ergebnis\FactoryBot\FixtureFactory;
 use Ergebnis\FactoryBot\Test\Fixture;
-use Faker\Generator;
 
 final class RepositoryDefinition implements Definition
 {
@@ -24,7 +23,7 @@ final class RepositoryDefinition implements Definition
     {
     }
 
-    public function accept(FixtureFactory $fixtureFactory, Generator $faker): void
+    public function accept(FixtureFactory $fixtureFactory): void
     {
         $fixtureFactory->define(Fixture\FixtureFactory\Entity\Repository::class);
     }